move FeatureExtractor to new featureEngine lib to avoid cyclic dependencies
