+++ src/aliceVision/CMakeLists.txt
@@ -15,6 +15,7 @@ if(ALICEVISION_BUILD_SFM)
   add_subdirectory(colorHarmonization)
   add_subdirectory(dataio)
   add_subdirectory(feature)
+  add_subdirectory(featureEngine)
   add_subdirectory(geometry)
   add_subdirectory(graph)
   add_subdirectory(gpu)
